Don't enable Lambda plugin in non-Lambda execution environments
The Lambda plugin currently enables itself in all environments where the
AWS_EXECUTION_ENV environment variable is set, regardless of value.

However, this variable is also set in other contexts, e.g. ECS. In those
contexts this results in the plugin overwriting the component and action
values set in any notice with nil (since no lambda function or handler
data is present in the environment).

This commit updates the requirement test to check the AWS_EXECUTION_ENV
value for the prefix specified in the AWS Lambda docs.[1]

  > AWS_EXECUTION_ENV –
  >   The runtime identifier, prefixed by AWS_Lambda_ — for example,
  >   AWS_Lambda_java8.

Since it's not completely clear what the capitalisation is I've opted to
make the regex comparison case insensitive; I think the risk of this
causing false positives is low.

The slightly odd implementation is to ensure that the result is always a
boolean as before; I've also used String#match instead of #match?
because the former is supported by older ruby versions.

[1] https://docs.aws.amazon.com/lambda/latest/dg/configuration-envvars.html